pyproject.toml
[project]
name = "pydanticai-api-template"
version = "0.1.0"
description = "A template for building FastAPI applications with PydanticAI."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "pydantic>=2.11.3,<3.0.0",
    "pydantic-ai[mcp]>=0.0.55",
    "openai>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.5.0",
    "shellingham>=1.5.0",
    "uvicorn==0.29.0",
    "websockets==12.0.0",
    "mcp>=0.5.0",
    "python-dotenv>=1.1.0",
    "typing-extensions>=4.7.0",
    "logfire>=0.16.0",
    "httpx>=0.24.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.11.0",
    "pre-commit>=3.3.3",
    "mypy>=1.5.1",
    "python-dotenv>=1.1.0",
    "tomli>=2.0.1",
    "pyyaml>=6.0.1",
    "promptfoo==0.1.0",
    "pytest>=8.0.0",
    "types-PyYAML>=6.0",
]
test = ["pytest>=7.4.0", "pytest-cov>=4.1.0", "pytest-asyncio>=0.23.0"]

[project.scripts]
pat = "pydanticai_api_template.cli:app"

[tool.hatch.build]
artifacts = ["src/**/*.py", "src/**/py.typed"]

[tool.hatch.build.targets.wheel]
packages = ["src/pydanticai_api_template"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_return_any = true
warn_unreachable = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "pydantic.*"
follow_imports = "skip"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "pydanticai_api_template.api.models"
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = "logfire.*"
ignore_missing_imports = true

# Relax decorator checks for tests and CLI due to pytest/typer interaction
[[tool.mypy.overrides]]
module = ["tests.*", "pydanticai_api_template.cli"]
disallow_untyped_decorators = false

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
